// src/ui/generated-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type Address = Node & {
  __typename?: 'Address';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  fullName?: Maybe<Scalars['String']>;
  company?: Maybe<Scalars['String']>;
  streetLine1: Scalars['String'];
  streetLine2?: Maybe<Scalars['String']>;
  city?: Maybe<Scalars['String']>;
  province?: Maybe<Scalars['String']>;
  postalCode?: Maybe<Scalars['String']>;
  country: Country;
  phoneNumber?: Maybe<Scalars['String']>;
  defaultShippingAddress?: Maybe<Scalars['Boolean']>;
  defaultBillingAddress?: Maybe<Scalars['Boolean']>;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type Channel = Node & {
  __typename?: 'Channel';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  code: Scalars['String'];
  token: Scalars['String'];
  defaultTaxZone?: Maybe<Zone>;
  defaultShippingZone?: Maybe<Zone>;
  defaultLanguageCode: LanguageCode;
  currencyCode: CurrencyCode;
  pricesIncludeTax: Scalars['Boolean'];
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type Nfe = Node & {
  __typename?: 'Nfe';
  id: Scalars['ID'];
  createdAt?: Maybe<Scalars['DateTime']>;
  updatedAt?: Maybe<Scalars['DateTime']>;
  state?: Maybe<Scalars['String']>;
  env: NfeEnv;
  series?: Maybe<Scalars['String']>;
  number?: Maybe<Scalars['String']>;
  receipt?: Maybe<Scalars['String']>;
  key?: Maybe<Scalars['String']>;
  xml?: Maybe<Scalars['String']>;
  danfe?: Maybe<Scalars['String']>;
  info: NfeInfo;
  recipient: NfeRecipient;
  payment: NfePayment;
  transport?: Maybe<NfeTransport>;
  exportation?: Maybe<NfeExportation>;
  references?: Maybe<NfeReferences>;
  company: NfeCompany;
  order: Order;
  lastOperation?: Maybe<NfeOperation>;
  operations?: Maybe<Array<Maybe<NfeOperation>>>;
  products?: Maybe<Array<NfeProduct>>;
};

export type NfeCompany = Node & {
  __typename?: 'NfeCompany';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  ownerName: Scalars['String'];
  corporateName: Scalars['String'];
  email: Scalars['String'];
  phone: Scalars['String'];
  tradeName?: Maybe<Scalars['String']>;
  accounting?: Maybe<Scalars['String']>;
  subdomain?: Maybe<Scalars['String']>;
  logoUrl?: Maybe<Scalars['String']>;
  unit: NfeCompanyUnit;
  cpf: Scalars['String'];
  cnpj: Scalars['String'];
  ie: Scalars['String'];
  address: NfeCompanyAddress;
  taxSettings?: Maybe<NfeCompanyTaxSettings>;
  certificate?: Maybe<NfeCompanyCertificate>;
  emitterCredentials?: Maybe<NfeCompanyEmitterCredentials>;
  channels?: Maybe<Array<Maybe<Channel>>>;
  nfes?: Maybe<Array<Maybe<Nfe>>>;
};

export type NfeCompanyAddress = {
  __typename?: 'NfeCompanyAddress';
  postcode: Scalars['String'];
  street: Scalars['String'];
  number: Scalars['String'];
  complement?: Maybe<Scalars['String']>;
  neighborhood: Scalars['String'];
  city: Scalars['String'];
  state: Scalars['String'];
};

export type NfeCompanyAddressInput = {
  postcode: Scalars['String'];
  street: Scalars['String'];
  number: Scalars['String'];
  complement: Scalars['String'];
  neighborhood: Scalars['String'];
  city: Scalars['String'];
  state: Scalars['String'];
};

export type NfeCompanyCertificate = {
  __typename?: 'NfeCompanyCertificate';
  base64?: Maybe<Scalars['String']>;
  password?: Maybe<Scalars['String']>;
};

export type NfeCompanyCertificateInput = {
  base64?: Maybe<Scalars['String']>;
  password?: Maybe<Scalars['String']>;
};

export type NfeCompanyEmitterCredentials = {
  __typename?: 'NfeCompanyEmitterCredentials';
  consumer_key?: Maybe<Scalars['String']>;
  consumer_secret?: Maybe<Scalars['String']>;
  access_token?: Maybe<Scalars['String']>;
  access_token_secret?: Maybe<Scalars['String']>;
};
// ...
export type NfeCompanyTaxSettings = {
  __typename?: 'NfeCompanyTaxSettings';
  ibpt?: Maybe<Scalars['String']>;
  informacoes_fisco?: Maybe<Scalars['String']>;
  orientacao_danfe?: Maybe<Scalars['String']>;
  nfe_serie?: Maybe<Scalars['Int']>;
  nfe_numero?: Maybe<Scalars['Int']>;
  nfe_numero_dev?: Maybe<Scalars['Int']>;
  nfce_serie?: Maybe<Scalars['Int']>;
  nfce_numero?: Maybe<Scalars['Int']>;
  nfce_numero_dev?: Maybe<Scalars['String']>;
  nfce_id_csc?: Maybe<Scalars['String']>;
  nfce_codigo_csc?: Maybe<Scalars['String']>;
  nfce_id_csc_dev?: Maybe<Scalars['String']>;
  nfce_codigo_csc_dev?: Maybe<Scalars['String']>;
};
// ...
export enum NfeCompanyUnit {
  Matriz = 'matriz',
  Filial = 'filial'
}
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export type Query = {
  __typename?: 'Query';
  administrators: AdministratorList;
  administrator?: Maybe<Administrator>;
  activeAdministrator?: Maybe<Administrator>;
  /** Get a list of Assets */
  assets: AssetList;
  /** Get a single Asset by id */
  asset?: Maybe<Asset>;
  me?: Maybe<CurrentUser>;
  channels: Array<Channel>;
  channel?: Maybe<Channel>;
  activeChannel: Channel;
  collections: CollectionList;
  /** Get a Collection either by id or slug. If neither id nor slug is specified, an error will result. */
  collection?: Maybe<Collection>;
  collectionFilters: Array<ConfigurableOperationDefinition>;
  /** Used for real-time previews of the contents of a Collection */
  previewCollectionVariants: ProductVariantList;
  countries: CountryList;
  country?: Maybe<Country>;
  customerGroups: CustomerGroupList;
  customerGroup?: Maybe<CustomerGroup>;
  customers: CustomerList;
  customer?: Maybe<Customer>;
  facets: FacetList;
  facet?: Maybe<Facet>;
  facetValues: FacetValueList;
  globalSettings: GlobalSettings;
  job?: Maybe<Job>;
  jobs: JobList;
  jobsById: Array<Job>;
  jobQueues: Array<JobQueue>;
  jobBufferSize: Array<JobBufferSize>;
  order?: Maybe<Order>;
  orders: OrderList;
  /** Returns a list of eligible shipping methods for the draft Order */
  eligibleShippingMethodsForDraftOrder: Array<ShippingMethodQuote>;
  paymentMethods: PaymentMethodList;
  paymentMethod?: Maybe<PaymentMethod>;
  paymentMethodEligibilityCheckers: Array<ConfigurableOperationDefinition>;
  paymentMethodHandlers: Array<ConfigurableOperationDefinition>;
  productOptionGroups: Array<ProductOptionGroup>;
  productOptionGroup?: Maybe<ProductOptionGroup>;
  search: SearchResponse;
  pendingSearchIndexUpdates: Scalars['Int'];
  /** List Products */
  products: ProductList;
  /** Get a Product either by id or slug. If neither id nor slug is specified, an error will result. */
  product?: Maybe<Product>;
  /** List ProductVariants either all or for the specific product. */
  productVariants: ProductVariantList;
  /** Get a ProductVariant by id */
  productVariant?: Maybe<ProductVariant>;
  promotion?: Maybe<Promotion>;
  promotions: PromotionList;
  promotionConditions: Array<ConfigurableOperationDefinition>;
  promotionActions: Array<ConfigurableOperationDefinition>;
  roles: RoleList;
  role?: Maybe<Role>;
  shippingMethods: ShippingMethodList;
  shippingMethod?: Maybe<ShippingMethod>;
  shippingEligibilityCheckers: Array<ConfigurableOperationDefinition>;
  shippingCalculators: Array<ConfigurableOperationDefinition>;
  fulfillmentHandlers: Array<ConfigurableOperationDefinition>;
  testShippingMethod: TestShippingMethodResult;
  testEligibleShippingMethods: Array<ShippingMethodQuote>;
  tag: Tag;
  tags: TagList;
  taxCategories: Array<TaxCategory>;
  taxCategory?: Maybe<TaxCategory>;
  taxRates: TaxRateList;
  taxRate?: Maybe<TaxRate>;
  zones: Array<Zone>;
  zone?: Maybe<Zone>;
  orderNfe?: Maybe<Nfe>;
  nfeCompany?: Maybe<NfeCompany>;
  nfeCompanyByChannel?: Maybe<NfeCompany>;
  nfeCompanies: NfeCompanyList;
  nfeSetting?: Maybe<NfeSetting>;
  nfeSettings: NfeSettingList;
  nfeSettingByTypeAndMethodId?: Maybe<NfeSetting>;
  platform?: Maybe<Platform>;
  platforms: PlatformList;
  platformResource?: Maybe<PlatformResource>;
  platformResources: PlatformResourceList;
  webhook?: Maybe<Scalars['String']>;
  packages: PackageList;
  package: Package;
};
// ...
  export type Fragment = NfeListItemFragment;
// ...
  export type Channels = NonNullable<(NonNullable<NfeCompanyListItemFragment['channels']>)[number]>;
// ...
export namespace NfeCompany {
  export type Fragment = NfeCompanyFragment;
  export type Address = (NonNullable<NfeCompanyFragment['address']>);
  export type TaxSettings = (NonNullable<NfeCompanyFragment['taxSettings']>);
  export type Certificate = (NonNullable<NfeCompanyFragment['certificate']>);
  export type EmitterCredentials = (NonNullable<NfeCompanyFragment['emitterCredentials']>);
  export type Channels = NonNullable<(NonNullable<NfeCompanyFragment['channels']>)[number]>;
}

export namespace NfeSetting {
  export type Fragment = NfeSettingFragment;
}
// ...
export namespace Nfe {
  export type Fragment = NfeFragment;
  export type Info = (NonNullable<NfeFragment['info']>);
  export type Recipient = (NonNullable<NfeFragment['recipient']>);
  export type Payment = (NonNullable<NfeFragment['payment']>);
  export type Fatura = (NonNullable<(NonNullable<NfeFragment['payment']>)['fatura']>);
  export type Parcelas = NonNullable<(NonNullable<(NonNullable<NfeFragment['payment']>)['parcelas']>)[number]>;
  export type Transport = (NonNullable<NfeFragment['transport']>);
  export type Exportation = (NonNullable<NfeFragment['exportation']>);
  export type References = (NonNullable<NfeFragment['references']>);
  export type Ecfs = NonNullable<(NonNullable<(NonNullable<NfeFragment['references']>)['ecfs']>)[number]>;
  export type Company = (NonNullable<NfeFragment['company']>);
  export type Order = (NonNullable<NfeFragment['order']>);
  export type LastOperation = (NonNullable<NfeFragment['lastOperation']>);
  export type Operations = NonNullable<(NonNullable<NfeFragment['operations']>)[number]>;
  export type Products = NonNullable<(NonNullable<NfeFragment['products']>)[number]>;
}
// ...
  export type Variables = CreateNfeCompanyMutationVariables;
// ...
export namespace GetNfeCompany {
  export type Variables = GetNfeCompanyQueryVariables;
  export type Query = GetNfeCompanyQuery;
  export type NfeCompany = (NonNullable<GetNfeCompanyQuery['nfeCompany']>);
}
// ...
export type ChannelFragmentFragment = (
  { __typename?: 'Channel' }
  & Pick<Channel, 'id' | 'code' | 'currencyCode'>
);

export type NfeCompanyFragment = (
  { __typename?: 'NfeCompany' }
  & Pick<NfeCompany, 'id' | 'createdAt' | 'updatedAt' | 'ownerName' | 'corporateName' | 'email' | 'phone' | 'tradeName' | 'accounting' | 'subdomain' | 'logoUrl' | 'cpf' | 'cnpj' | 'ie' | 'unit'>
  & { address: (
    { __typename?: 'NfeCompanyAddress' }
    & Pick<NfeCompanyAddress, 'postcode' | 'street' | 'number' | 'complement' | 'neighborhood' | 'city' | 'state'>
  ), taxSettings?: Maybe<(
    { __typename?: 'NfeCompanyTaxSettings' }
    & Pick<NfeCompanyTaxSettings, 'ibpt' | 'informacoes_fisco' | 'orientacao_danfe' | 'nfe_serie' | 'nfe_numero' | 'nfe_numero_dev' | 'nfce_serie' | 'nfce_numero' | 'nfce_numero_dev' | 'nfce_id_csc' | 'nfce_codigo_csc' | 'nfce_id_csc_dev' | 'nfce_codigo_csc_dev'>
  )>, certificate?: Maybe<(
    { __typename?: 'NfeCompanyCertificate' }
    & Pick<NfeCompanyCertificate, 'base64' | 'password'>
  )>, emitterCredentials?: Maybe<(
    { __typename?: 'NfeCompanyEmitterCredentials' }
    & Pick<NfeCompanyEmitterCredentials, 'consumer_key' | 'consumer_secret' | 'access_token' | 'access_token_secret'>
  )>, channels?: Maybe<Array<Maybe<(
    { __typename?: 'Channel' }
    & ChannelFragmentFragment
  )>>> }
);
// ...
export type CreateNfeCompanyMutationVariables = Exact<{
  input: CreateNfeCompanyInput;
}>;
// ...
export type GetNfeCompanyQueryVariables = Exact<{
  id: Scalars['ID'];
}>;


export type GetNfeCompanyQuery = (
  { __typename?: 'Query' }
  & { nfeCompany?: Maybe<(
    { __typename?: 'NfeCompany' }
    & NfeCompanyFragment
  )> }
);